Move getLogger and NameUtils out of kotlin package.
