+++ embabel-common-ai/src/main/kotlin/com/embabel/common/ai/model/ApplicationPropertiesModelProvider.kt
@@ -15,7 +15,7 @@
  */
 package com.embabel.common.ai.model
 
-import com.embabel.common.util.kotlin.loggerFor
+import com.embabel.common.util.loggerFor
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.validation.annotation.Validated
