+++ embabel-common-util/src/main/kotlin/com/embabel/common/util/NamingUtils.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.embabel.common.util.kotlin
+package com.embabel.common.util
 
 
 /**
